refactor: move index access check to a utils class
